requirements.txt
flask==3.1.0
anthropic==0.43.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
